pyproject.toml
[project]
name = "application"
version = "0.1.0"
description = ""
authors = [
    {name = "Leandro",email = "leandrohenriquesc@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg (>=0.31.0,<0.32.0)",
    "tortoise-orm (>=1.1.5,<2.0.0)",
    "sanic (>=25.12.0,<26.0.0)",
    "psycopg (>=3.3.3,<4.0.0)",
    "pydantic (>=2.12.5,<3.0.0)"
]

[tool.poetry]
packages = [{include = "application", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
